feat: enhance product details retrieval and category management
- Introduced `get_product_details` function to fetch comprehensive product information, including price and stock.
- Updated the ChatService to utilize the new product details function for improved customer interactions.
- Enhanced product and variant models to support floating-point pricing, ensuring accurate financial representation.
- Modified frontend components to enforce category selection during product creation, improving user experience.
- Updated database schema to reflect changes in product pricing and category relationships.

+++ apps/go-core/internal/models/product_variant.go
@@ -51,10 +51,10 @@ type ProductVariant struct {
 	// structured option values — { "color": "red", "size": "M" }
 	Attributes json.RawMessage `db:"attributes" json:"attributes,omitempty"`
 
-	// Pricing (cents) — null means inherit the parent product's value
-	Price     *int `db:"price"      json:"price,omitempty"`
-	CostPrice *int `db:"cost_price" json:"cost_price,omitempty"` // never exposed to customer
-	Discount  *int `db:"discount"   json:"discount,omitempty"`
+	// Pricing — null means inherit the parent product's value
+	Price     *float64 `db:"price"      json:"price,omitempty"`
+	CostPrice *float64 `db:"cost_price" json:"cost_price,omitempty"` // never exposed to customer
+	Discount  *int     `db:"discount"   json:"discount,omitempty"`
 
 	// Inventory
 	StockQty          int  `db:"stock_qty"           json:"stock_qty"`
@@ -78,9 +78,9 @@ type CreateProductVariantInput struct {
 
 	Attributes json.RawMessage `json:"attributes"`
 
-	Price     *int `json:"price"      validate:"omitempty,gt=0"`
-	CostPrice *int `json:"cost_price" validate:"omitempty,gt=0"`
-	Discount  *int `json:"discount"   validate:"omitempty,min=0,max=100"`
+	Price     *float64 `json:"price"      validate:"omitempty,gt=0"`
+	CostPrice *float64 `json:"cost_price" validate:"omitempty,gt=0"`
+	Discount  *int     `json:"discount"   validate:"omitempty,min=0,max=100"`
 
 	StockQty          int  `json:"stock_qty"           validate:"min=0"`
 	LowStockThreshold *int `json:"low_stock_threshold" validate:"omitempty,min=0"`
@@ -97,9 +97,9 @@ type UpdateProductVariantInput struct {
 
 	Attributes json.RawMessage `json:"attributes"`
 
-	Price     *int `json:"price"      validate:"omitempty,gt=0"`
-	CostPrice *int `json:"cost_price" validate:"omitempty,gt=0"`
-	Discount  *int `json:"discount"   validate:"omitempty,min=0,max=100"`
+	Price     *float64 `json:"price"      validate:"omitempty,gt=0"`
+	CostPrice *float64 `json:"cost_price" validate:"omitempty,gt=0"`
+	Discount  *int     `json:"discount"   validate:"omitempty,min=0,max=100"`
 
 	StockQty          *int `json:"stock_qty"           validate:"omitempty,min=0"`
 	LowStockThreshold *int `json:"low_stock_threshold" validate:"omitempty,min=0"`
